# src/swagger.api.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is for meta schema development",
    "version": "1.0.0",
    "title": "Meta Schema Creator",
    "contact": {
      "email": "arvindrawat111@gmail.com"
    }
  },
  "host": "localhost:8080/",
  "basePath": "api/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/auto": {
      "get": {
        "tags": [
          "AutoGenerator"
        ],
        "summary": "Get Meta module by id",
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "post": {
        "tags": [
          "AutoGenerator"
        ],
        "summary": "Create meta module",
        "description": "Creat new meta module.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created meta module object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/auto/{pluralmodulename}/{id}": {
      "get": {
        "tags": [
          "AutoGenerator"
        ],
        "summary": "Get Item by id",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pluralmodulename",
            "description": "Plural module name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "description": "Item id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "put": {
        "tags": [
          "AutoGenerator"
        ],
        "summary": "Update Item",
        "description": "Update Item.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pluralmodulename",
            "description": "Plural module name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "description": "Item id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid id supplied"
          },
          "404": {
            "description": "Item not found"
          }
        }
      },
      "delete": {
        "tags": [
          "AutoGenerator"
        ],
        "summary": "Delete Item",
        "description": "Delete Item.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pluralmodulename",
            "description": "Plural module name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "description": "Item id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid Item supplied"
          },
          "404": {
            "description": "Item not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
...
}